renamed cactus-reference-align to cactus-refmap; replaced subprocess calls to cactus_call
