--- src/cactus/reference_align/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-
-#Copyright (C) 2009-2011 by Benedict Paten (benedictpaten@gmail.com)
-#
-#Released under the MIT license, see LICENSE.txt
